requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
certifi==2019.9.11
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.4.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
packaging==20.9
pipenv==2018.11.26
prompt-toolkit==3.0.18
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1
